CP_Image_DrawSubImageAdvanced
Added CP_Image_DrawSubImageAdvanced. This only took adding another function that used an already existing parameter. I created the quad-justin.png file to test and make sure that the sub-image was working. I modified the i_test state so that if you wanted to make it so that if you press space it toggles between a full justin face and a quarter of the 2x2 justin image. There is no visible change but that means that it's working right.